Main class for running program factored out
